move sequence op unittests to a separate folder test=develop
