+++ python/paddle/fluid/tests/unittests/sequence/test_seq_pool.py
