+++ python/paddle/fluid/tests/unittests/sequence/test_sequence_enumerate_op.py
