+++ python/paddle/fluid/tests/unittests/sequence/test_sequence_erase_op.py
