+++ python/paddle/fluid/tests/unittests/sequence/test_sequence_expand.py
