+++ python/paddle/fluid/tests/unittests/sequence/test_sequence_expand_as.py
