+++ python/paddle/fluid/tests/unittests/sequence/test_sequence_scatter_op.py
