+++ python/paddle/fluid/tests/unittests/sequence/test_sequence_softmax_op.py
