+++ python/paddle/fluid/tests/unittests/test_fusion_seqconv_eltadd_relu_op.py
@@ -18,7 +18,7 @@ import unittest
 import numpy as np
 import random
 from op_test import OpTest
-from test_seq_conv import seqconv
+from sequence.test_seq_conv import seqconv
 
 
 class TestSeqConvEltAddRelu(OpTest):
