+++ python/paddle/fluid/tests/unittests/test_fusion_seqpool_concat_op.py
@@ -18,7 +18,7 @@ import unittest
 import numpy as np
 from op_test import OpTest
 from test_reorder_lod_tensor import convert_to_offset
-from test_seq_pool import compute_seqpool_sum, compute_seqpool_avg, compute_seqpool_sqrt
+from sequence.test_seq_pool import compute_seqpool_sum, compute_seqpool_avg, compute_seqpool_sqrt
 
 
 class TestFusionSeqPoolConcatOp(OpTest):
